Clear warnings

+++ src/syscall.c
@@ -28,7 +28,7 @@ int open(const char *pathname, int flags, ...)
 {
     int ret;
 
-    ret = syscall(SYS_OPEN, pathname, flags, 0, 0);
+    ret = syscall(SYS_OPEN, (uint64_t)pathname, flags, 0, 0);
 
     return ret;
 }
@@ -46,7 +46,7 @@ ssize_t read(int fd, void *buf, size_t count)
 {
     int ret;
 
-    ret = syscall(SYS_READ, fd, buf, count, 0);
+    ret = syscall(SYS_READ, fd, (uint64_t)buf, count, 0);
 
     return ret;
 }
@@ -55,7 +55,7 @@ ssize_t write(int fd, const void *buf, size_t count)
 {
     int ret;
 
-    ret = syscall(SYS_WRITE, fd, buf, count, 0);
+    ret = syscall(SYS_WRITE, fd, (uint64_t)buf, count, 0);
 
     return ret;
 }
@@ -64,7 +64,7 @@ int creat(const char *pathname, mode_t mode)
 {
     int ret;
 
-    ret = syscall(SYS_CREATE, pathname, mode, 0, 0);
+    ret = syscall(SYS_CREATE, (uint64_t)pathname, mode, 0, 0);
 
     return ret;
 }
